[Decomp][LFG] Cohesion split: extract queue + proposal/role flow to LFGMgr*.cpp
